Add mock.ElementsMatch argument matcher
- Refactor argumentMatcher to be an interface.
- Refactor move code common to assert and mock into internal package.
